requirements.txt

streamlit==1.22.0
tensorflow-cpu==2.9.0
numpy==1.21.6
